Corrections to Slovenian translation.
According to the Slovenian Terminological Dictionary of Informatics
(http://www.islovar.org), the word "brkljalnik" ("browser") is deprecated in
favour of "brskalnik".

Corrected the spelling of "Berenikini kodri" (Coma Berenices).

Capella: replaced the inappropriate direct translation from Latin ("Kozica",
literally "a small goat") with the name of the star as is actually used in
Slovenian ("Kapela").